fix(compiler): Replace TypeScript compiler
Remove call to opn so that the default browser will not open when running the
dev server in production or development.

Replace TypeScript compiler with Babel so that CommonJS modules with a default
export can be imported using ES2015 module syntax.

BREAKING CHANGE: async/await will no longer down comiler to ES5.